inverted depth calc

+++ src/uci/commands.rs
@@ -34,7 +34,7 @@ pub enum UciOutput {
 
 #[derive(Debug)]
 pub struct Info {
-    pub depth: u8,
+    pub depth: u32,
     pub nodes: u32,
     pub nodes_per_second: u32,
     pub time: u32,
